+++ src/layouts/dashboard/DashboardSidebar.tsx
@@ -1,10 +1,10 @@
 import React from "react";
 import PropTypes from 'prop-types';
-import { useEffect } from 'react';
-import { Link as RouterLink, useLocation } from 'react-router-dom';
+import {useEffect} from 'react';
+import {Link as RouterLink, useLocation} from 'react-router-dom';
 // material
-import { styled } from '@mui/material/styles';
-import { Box, Link, Button, Drawer, Typography, Avatar, Stack } from '@mui/material';
+import {styled} from '@mui/material/styles';
+import {Box, Link, Button, Drawer, Typography, Avatar, Stack} from '@mui/material';
 // mock
 import account from '../../_mock/account';
 // hooks
@@ -12,7 +12,7 @@ import useResponsive from '../../hooks/useResponsive';
 // components
 import Logo from '../../components/Logo';
 import Scrollbar from '../../components/Scrollbar';
-import NavSection from '../../components/NavSection';
+import NavSection from '../../components/NavSection.tsx';
 //
 import navConfig from './NavConfig';
 
@@ -25,7 +25,7 @@ const DRAWER_WIDTH = 280;
 //   },
 // }));
 
-const AccountStyle = styled('div')(({ theme }) => ({
+const AccountStyle = styled('div')(({theme}) => ({
   display: 'flex',
   alignItems: 'center',
   padding: theme.spacing(2, 2.5),
@@ -40,8 +40,8 @@ DashboardSidebar.propTypes = {
   onCloseSidebar: PropTypes.func,
 };
 
-export default function DashboardSidebar({ isOpenSidebar, onCloseSidebar }) {
-  const { pathname } = useLocation();
+export default function DashboardSidebar({isOpenSidebar, onCloseSidebar}) {
+  const {pathname} = useLocation();
 
   const isDesktop = useResponsive('up', 'lg');
 
@@ -56,22 +56,22 @@ export default function DashboardSidebar({ isOpenSidebar, onCloseSidebar }) {
     <Scrollbar
       sx={{
         height: 1,
-        '& .simplebar-content': { height: 1, display: 'flex', flexDirection: 'column' },
+        '& .simplebar-content': {height: 1, display: 'flex', flexDirection: 'column'},
       }}
     >
-      <Box sx={{ px: 2.5, py: 3, display: 'inline-flex' }}>
-        <Logo />
+      <Box sx={{px: 2.5, py: 3, display: 'inline-flex'}}>
+        <Logo/>
       </Box>
 
-      <Box sx={{ mb: 5, mx: 2.5 }}>
+      <Box sx={{mb: 5, mx: 2.5}}>
         <Link underline="none" component={RouterLink} to="#">
           <AccountStyle>
-            <Avatar src={account.photoURL} alt="photoURL" />
-            <Box sx={{ ml: 2 }}>
-              <Typography variant="subtitle2" sx={{ color: 'text.primary' }}>
+            <Avatar src={account.photoURL} alt="photoURL"/>
+            <Box sx={{ml: 2}}>
+              <Typography variant="subtitle2" sx={{color: 'text.primary'}}>
                 {account.displayName}
               </Typography>
-              <Typography variant="body2" sx={{ color: 'text.secondary' }}>
+              <Typography variant="body2" sx={{color: 'text.secondary'}}>
                 {account.role}
               </Typography>
             </Box>
@@ -79,32 +79,10 @@ export default function DashboardSidebar({ isOpenSidebar, onCloseSidebar }) {
         </Link>
       </Box>
 
-      <NavSection navConfig={navConfig} />
+      <NavSection navConfig={navConfig}/>
 
-      <Box sx={{ flexGrow: 1 }} />
+      <Box sx={{flexGrow: 1}}/>
 
-      <Box sx={{ px: 2.5, pb: 3, mt: 10 }}>
-        <Stack alignItems="center" spacing={3} sx={{ pt: 5, borderRadius: 2, position: 'relative' }}>
-          <Box
-            component="img"
-            src="/static/illustrations/illustration_avatar.png"
-            sx={{ width: 100, position: 'absolute', top: -50 }}
-          />
-
-          <Box sx={{ textAlign: 'center' }}>
-            <Typography gutterBottom variant="h6">
-              Get more?
-            </Typography>
-            <Typography variant="body2" sx={{ color: 'text.secondary' }}>
-              From only $69
-            </Typography>
-          </Box>
-
-          <Button href="https://material-ui.com/store/items/minimal-dashboard/" target="_blank" variant="contained">
-            Upgrade to Pro
-          </Button>
-        </Stack>
-      </Box>
     </Scrollbar>
   );
 
@@ -115,7 +93,7 @@ export default function DashboardSidebar({ isOpenSidebar, onCloseSidebar }) {
           open={isOpenSidebar}
           onClose={onCloseSidebar}
           PaperProps={{
-            sx: { width: DRAWER_WIDTH },
+            sx: {width: DRAWER_WIDTH},
           }}
         >
           {renderContent}