+++ src/sections/@dashboard/app/AppCurrentSubject.tsx
@@ -1,0 +1,71 @@
+import React, {FC} from "react";
+import PropTypes from 'prop-types';
+import merge from 'lodash/merge';
+import ReactApexChart from 'react-apexcharts';
+// @mui
+import { styled } from '@mui/material/styles';
+import { Card, CardHeader } from '@mui/material';
+// components
+import { BaseOptionChart } from '../../../components/chart';
+
+// ----------------------------------------------------------------------
+
+const CHART_HEIGHT = 392;
+
+const LEGEND_HEIGHT = 72;
+
+const ChartWrapperStyle = styled('div')(({ theme }) => ({
+  height: CHART_HEIGHT,
+  marginTop: theme.spacing(2),
+  '& .apexcharts-canvas svg': {
+    height: CHART_HEIGHT,
+  },
+  '& .apexcharts-canvas svg,.apexcharts-canvas foreignObject': {
+    overflow: 'visible',
+  },
+  '& .apexcharts-legend': {
+    height: LEGEND_HEIGHT,
+    alignContent: 'center',
+    position: 'relative !important',
+    borderTop: `solid 1px ${theme.palette.divider}`,
+    top: `calc(${CHART_HEIGHT - LEGEND_HEIGHT}px) !important`,
+  },
+}));
+
+// ----------------------------------------------------------------------
+
+interface IAppCurrentSubject {
+  title: string
+  subheader?: string
+  chartData: Array<any>
+  chartColors: Array<any>,
+  chartLabels: Array<any>,
+}
+
+const AppCurrentSubject:FC<IAppCurrentSubject> = ({ title, subheader, chartData, chartColors, chartLabels, ...other }) => {
+  const chartOptions = merge(BaseOptionChart(), {
+    stroke: { width: 2 },
+    fill: { opacity: 0.48 },
+    legend: { floating: true, horizontalAlign: 'center' },
+    xaxis: {
+      categories: chartLabels,
+      labels: {
+        style: {
+          colors: chartColors,
+        },
+      },
+    },
+  });
+
+  return (
+    <Card {...other}>
+      <CardHeader title={title} subheader={subheader} />
+
+      <ChartWrapperStyle dir="ltr">
+        <ReactApexChart type="radar" series={chartData} options={chartOptions as any} height={340} />
+      </ChartWrapperStyle>
+    </Card>
+  );
+}
+
+export default AppCurrentSubject